refactor(lib): extract affiliates-data + redirect-networks from affiliates.js (#826 PR1)
Split the 1253-line affiliates.js monolith into three focused modules:

- src/lib/affiliates-data.js (778 lines) — static tracking-param dataset:
  TRACKING_PARAMS, TRACKING_PREFIXES, TRACKING_PARAM_CATEGORIES

- src/lib/redirect-networks.js (287 lines) — redirect-network table + lookups:
  REDIRECT_NETWORK_PATTERNS, getRedirectNetworkPatterns,
  getRedirectNetworkForRedirectHost, getLandingParamsForReferrer

- src/lib/affiliates.js (237 lines) — affiliate-program registry (domain b)
  + re-exports from the two new modules so ALL existing importers continue
  to work unchanged without any forced churn.

Acyclicity: data modules import nothing from affiliates.js.
Module-boundary guard: tests/unit/module-boundary-826.test.mjs added.

Source-path-coupled consumers updated (read file as text, not just import):
- tools/generate-rules.mjs: AFFILIATES_PATH → affiliates-data.js
- tools/add-rule.mjs: AFFILIATES_PATH → affiliates-data.js
- tests/unit/docs-claims.test.mjs: source-read path → affiliates-data.js
- tests/unit/rules-manifest-sync.test.mjs: source-read path → affiliates-data.js

Parity: TRACKING_PARAMS.length=448, AFFILIATE_PATTERNS.length=6 (unchanged).
Gate: npm test (4732 pass), lint:js (clean), typecheck (clean),
      test:integration:stub (196 pass), build:content (bundle regenerated).
