+++ tools/add-rule.mjs
@@ -14,9 +14,10 @@
  *
  * What the script changes (in this order):
  *
- *   1. src/lib/affiliates.js
+ *   1. src/lib/affiliates-data.js  (#826: moved from affiliates.js)
  *      - Appends the param to the flat `TRACKING_PARAMS` array
  *      - Appends the param to the matching `TRACKING_PARAM_CATEGORIES.<cat>.params` array
+ *      (affiliates.js re-exports both so all existing importers are unaffected)
  *   2. src/rules/tracking-params.json
  *      - Regenerated by `npm run compile:rules` so DNR stays in sync
  *   2b. src/rules/rules-manifest.json
@@ -53,10 +54,13 @@ import { fileURLToPath } from "node:url";
 
 const __dirname = dirname(fileURLToPath(import.meta.url));
 const REPO_ROOT = join(__dirname, "..");
-const AFFILIATES_PATH = join(REPO_ROOT, "src", "lib", "affiliates.js");
+// #826: TRACKING_PARAMS and TRACKING_PARAM_CATEGORIES moved to affiliates-data.js.
+// This tool mutates the data module directly; affiliates.js re-exports both so
+// all existing importers continue to work unchanged.
+const AFFILIATES_PATH = join(REPO_ROOT, "src", "lib", "affiliates-data.js");
 const TEST_PATH = join(REPO_ROOT, "tests", "unit", "cleaner-add-rule-regression.test.mjs");
 
-// Must match the keys of TRACKING_PARAM_CATEGORIES in src/lib/affiliates.js.
+// Must match the keys of TRACKING_PARAM_CATEGORIES in src/lib/affiliates-data.js.
 // Edits to that map must update this list (or the validation flips false).
 const CATEGORIES = ["utm", "ads", "email", "social", "platform_noise", "generic"];
 
@@ -124,9 +128,9 @@ function appendToTrackingParams(source, name, contextComment) {
   // Already-stripped detection: the literal `"name"` token appears
   // inside the TRACKING_PARAMS block somewhere.
   const exportIdx = source.indexOf("export const TRACKING_PARAMS = [");
-  if (exportIdx === -1) throw new Error("affiliates.js: TRACKING_PARAMS export not found");
+  if (exportIdx === -1) throw new Error("affiliates-data.js: TRACKING_PARAMS export not found");
   const closeIdx = source.indexOf("\n];", exportIdx);
-  if (closeIdx === -1) throw new Error("affiliates.js: TRACKING_PARAMS close `];` not found");
+  if (closeIdx === -1) throw new Error("affiliates-data.js: TRACKING_PARAMS close `];` not found");
   const block = source.slice(exportIdx, closeIdx);
   const tokenRe = new RegExp(`(^|\\W)"${escapeRegExp(name)}"`);
   if (tokenRe.test(block)) {
@@ -146,11 +150,11 @@ function appendToTrackingParams(source, name, contextComment) {
 function appendToCategory(source, category, name) {
   // Locate `<category>: {` then the `params: [` inside that object.
   const catIdx = source.indexOf(`\n  ${category}: {`);
-  if (catIdx === -1) throw new Error(`affiliates.js: TRACKING_PARAM_CATEGORIES.${category} not found`);
+  if (catIdx === -1) throw new Error(`affiliates-data.js: TRACKING_PARAM_CATEGORIES.${category} not found`);
   const paramsIdx = source.indexOf("params: [", catIdx);
-  if (paramsIdx === -1) throw new Error(`affiliates.js: ${category}.params not found`);
+  if (paramsIdx === -1) throw new Error(`affiliates-data.js: ${category}.params not found`);
   const closeIdx = source.indexOf("\n    ],", paramsIdx);
-  if (closeIdx === -1) throw new Error(`affiliates.js: ${category}.params close \`],\` not found`);
+  if (closeIdx === -1) throw new Error(`affiliates-data.js: ${category}.params close \`],\` not found`);
   const block = source.slice(paramsIdx, closeIdx);
   const tokenRe = new RegExp(`(^|\\W)"${escapeRegExp(name)}"`);
   if (tokenRe.test(block)) {
@@ -238,8 +242,9 @@ async function main() {
   const flags = parseFlags(process.argv.slice(2));
   const { name, category, source } = await gatherInput(flags);
 
-  // Mutate affiliates.js in two steps so partial failure doesn't leave
-  // the file half-written.
+  // Mutate affiliates-data.js in two steps so partial failure doesn't leave
+  // the file half-written. affiliates.js re-exports both symbols so all
+  // existing importers pick up the change transparently.
   let affSrc = readFileSync(AFFILIATES_PATH, "utf8");
   const r1 = appendToTrackingParams(affSrc, name, source);
   if (r1.changed) affSrc = r1.next;
@@ -247,9 +252,9 @@ async function main() {
   if (r2.changed) affSrc = r2.next;
   if (r1.changed || r2.changed) {
     writeFileSync(AFFILIATES_PATH, affSrc);
-    process.stdout.write(`✓ affiliates.js: TRACKING_PARAMS=${r1.changed ? "added" : "skipped"}, ${category}.params=${r2.changed ? "added" : "skipped"}\n`);
+    process.stdout.write(`✓ affiliates-data.js: TRACKING_PARAMS=${r1.changed ? "added" : "skipped"}, ${category}.params=${r2.changed ? "added" : "skipped"}\n`);
   } else {
-    process.stdout.write(`✓ affiliates.js: param already present in both lists, skipping\n`);
+    process.stdout.write(`✓ affiliates-data.js: param already present in both lists, skipping\n`);
   }
 
   ensureTestFile();
@@ -269,7 +274,7 @@ async function main() {
   // Stage but DO NOT commit.
   process.stdout.write(`\n→ git add (staging the diff for review)\n`);
   execSync(
-    "git add src/lib/affiliates.js src/rules/tracking-params.json src/rules/rules-manifest.json src/content/cleaner-bundle.js " + TEST_PATH,
+    "git add src/lib/affiliates-data.js src/rules/tracking-params.json src/rules/rules-manifest.json src/content/cleaner-bundle.js " + TEST_PATH,
     opts,
   );
